requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
aiosqlite
python-dotenv
httpx
ccxt
pandas
apscheduler
